security: tf_driver: update with latest TL drop
This change fixes abnormal signal handling in tf driver.
Tf_daemon(user thread) is waiting for the signal when coming out of LP0,
but tf_daemon never gets the signal because tf driver can't handle the
signal properly. If there's a pending signal when coming out of LP0,
this fix clears the pending signal.

Bug 1244750
Bug 1309812

Change-Id: I7849866c7993af8716c17f6b7a06692271334664
Signed-off-by: Hyung Taek Ryoo <hryoo@nvidia.com>
Reviewed-on: http://git-master/r/253586
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Jun Yan <juyan@nvidia.com>
Tested-by: Jun Yan <juyan@nvidia.com>
Reviewed-by: Anshul Jain (SW) <anshulj@nvidia.com>